Move ObjectMapperUtil para o modúlo core.
