+++ core/src/test/java/br/com/alc/ecommerce/checkout/core/util/AllUtilUnitTests.java
@@ -7,7 +7,8 @@ import org.junit.platform.suite.api.SuiteDisplayName;
 @Suite
 @SelectClasses({
         DateUtilTest.class,
-        EnumUtilTest.class
+        EnumUtilTest.class,
+        ObjectMapperUtilTest.class
 })
 @SuppressWarnings("squid:S2187")
 @SuiteDisplayName("Suite that gathers all unit tests of the Util package")
