+++ infrastructure/src/main/java/br/com/alc/ecommerce/checkout/infrastructure/adapter/input/impl/SaleProcessorInAdapterImpl.java
@@ -9,7 +9,7 @@ import lombok.extern.log4j.Log4j2;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Component;
 
-import static br.com.alc.ecommerce.checkout.infrastructure.util.ObjectMapperUtil.generateJson;
+import static br.com.alc.ecommerce.checkout.core.util.ObjectMapperUtil.generateJson;
 
 @Log4j2
 @Component
