+++ infrastructure/src/main/java/br/com/alc/ecommerce/checkout/infrastructure/adapter/output/SaleOrderInserterOutPortImpl.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Component;
 
 import java.util.Optional;
 
-import static br.com.alc.ecommerce.checkout.infrastructure.util.ObjectMapperUtil.generateJson;
+import static br.com.alc.ecommerce.checkout.core.util.ObjectMapperUtil.generateJson;
 
 @Log4j2
 @Component
