+++ infrastructure/src/main/java/br/com/alc/ecommerce/checkout/infrastructure/messaging/listener/SaleListener.java
@@ -7,7 +7,7 @@ import lombok.extern.log4j.Log4j2;
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.stereotype.Service;
 
-import static br.com.alc.ecommerce.checkout.infrastructure.util.ObjectMapperUtil.generateJson;
+import static br.com.alc.ecommerce.checkout.core.util.ObjectMapperUtil.generateJson;
 import static org.apache.commons.lang3.exception.ExceptionUtils.getMessage;
 
 @Log4j2
