+++ infrastructure/src/main/java/br/com/alc/ecommerce/checkout/infrastructure/messaging/producer/impl/MessagingProducerImpl.java
@@ -9,7 +9,7 @@ import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
 
-import static br.com.alc.ecommerce.checkout.infrastructure.util.ObjectMapperUtil.generateJson;
+import static br.com.alc.ecommerce.checkout.core.util.ObjectMapperUtil.generateJson;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.springframework.amqp.core.MessageProperties.CONTENT_TYPE_JSON;
