+++ infrastructure/src/test/java/br/com/alc/ecommerce/checkout/infrastructure/UnitTests.java
@@ -2,7 +2,6 @@ package br.com.alc.ecommerce.checkout.infrastructure;
 
 import br.com.alc.ecommerce.checkout.infrastructure.config.BeanConfigurationTest;
 import br.com.alc.ecommerce.checkout.infrastructure.messaging.producer.impl.MessagingProducerImplTest;
-import br.com.alc.ecommerce.checkout.infrastructure.util.ObjectMapperUtilTest;
 import org.junit.platform.suite.api.SelectClasses;
 import org.junit.platform.suite.api.Suite;
 import org.junit.platform.suite.api.SuiteDisplayName;
@@ -10,8 +9,7 @@ import org.junit.platform.suite.api.SuiteDisplayName;
 @Suite
 @SelectClasses({
         BeanConfigurationTest.class,
-        MessagingProducerImplTest.class,
-        ObjectMapperUtilTest.class
+        MessagingProducerImplTest.class
 })
 @SuppressWarnings("squid:S2187")
 @SuiteDisplayName(" Suite that gathers all Unit tests")
